Single Action Controller strategy init
