+++ models/contacts.js
@@ -1,7 +1,6 @@
 import mongoose from "mongoose";
 
 const Schema = mongoose.Schema;
-
 const contact = new Schema(
   {
     name: {
